FitnessPlatform/requirements.txt
# Основные зависимости Flask
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.4
Flask-Mail==0.9.1
Flask-CORS==4.0.0
Flask-Session==0.5.0
Flask-Moment==1.0.5
Flask-RESTX==1.1.0

# Работа с данными и БД
Werkzeug==2.3.6
WTForms==3.0.1
python-dotenv==1.0.0
SQLAlchemy==1.4.46
alembic==1.11.1
marshmallow==3.20.1
email-validator==2.1.0
pandas==2.1.3
numpy==1.24.3

# Безопасность
bcrypt==4.0.1
cryptography==41.0.3

# Фронтенд и визуализация
chart.js==4.4.0
plotly==5.17.0
reportlab==4.0.4

# Утилиты и обработка
requests==2.31.0
Pillow==10.0.0
python-dateutil==2.8.2
pytz==2023.3
redis==4.6.0
celery==5.3.4
structlog==23.1.0

# Разработка и тестирование
blinker==1.6.2
click==8.1.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pytest==7.4.2
pytest-flask==1.2.0
Faker==19.6.2
factory-boy==3.3.0
coverage==7.3.2

# Фронтенд библиотеки (используются через CDN, но для сборки могут понадобиться)
# Эти зависимости нужны, если будете собирать статику локально
# bootstrap==5.3.2
# fontawesomefree==6.4.2
# fullcalendar==5.11.3
# flatpickr==4.6.13